SCRUM-34 refactor: extract types in separate files
