+++ src/store/services/postsApi/postsApi.ts
@@ -1,39 +1,6 @@
 import { createApi } from "@reduxjs/toolkit/query/react";
 import { baseQueryUpdateToken } from "store/services/baseApi/baseApi";
-
-export type PostsWithMeta = {
-  pageSize: number;
-  totalCount: number;
-  notReadCount: number;
-  items: Post[];
-};
-export type Post = {
-  id: number;
-  userName: string;
-  description: string;
-  location: string;
-  images: Image[];
-  createdAt: string;
-  updatedAt: string;
-  ownerId: number;
-  avatarOwner: string;
-  owner: PostOwner;
-  likesCount: number;
-  isLiked: boolean;
-  avatarWhoLikes: boolean;
-};
-export type Image = {
-  url: string;
-  width: number;
-  height: number;
-  fileSize: number;
-  createdAt: string;
-  uploadId: string;
-};
-export type PostOwner = {
-  firstName: string;
-  lastName: string;
-};
+import { PostsWithMeta } from "store/services/postsApi/postsApi.types";
 
 export const postsApi = createApi({
   reducerPath: "postsApi",
